requirements.txt
Django==1.5.4
djangorestframework==2.3.12
gunicorn==0.17.2